Add files via upload
The actual code used to plot the original TMRCA graphs with their metrics. Though this representation comprises panels with all pairs on top of concordant pairs and discordant pairs alone, I decided to edit different of these graphs to obtain the figure in my report, with only the important information I want to highlight.